use public access modifier for mscv correct linkage.
private static fileds are mangled as C and public static fields are mangled as S on MSVC. since fields with no access modifeiers in C++ classes are private by default, it affects the mangling of the fields on windows so need for public there
also fields that are private on the D side should match on C++ side. Posix doesn't really care about those but msvc does.

match private with private for msvc